Get the current time with bash instead of groovy
The script was grabbing the time using the System.currentTimeSeconds()
function which fails to run in a sandbox environment (used when
running the pipeline from a Jenkinsfile). This script changes the
code to grab the current time in seconds using a shell script and
saving the output to a variable.